sample-mcp-server/requirements.txt
mcp[cli]
httpx
python-dotenv
html2text
requests
beautifulsoup4
